+++ harmony_browse_image_generator/browse.py
@@ -18,12 +18,19 @@ from PIL import Image
 from rasterio.io import DatasetReader
 from rasterio.plot import reshape_as_image, reshape_as_raster
 from rasterio.warp import Resampling, reproject
+from rioxarray import open_rasterio
+from xarray import DataArray
 
-from harmony_browse_image_generator.color_utility import get_color_palette
-from harmony_browse_image_generator.exceptions import (
-    HyBIGError,
+from harmony_browse_image_generator.color_utility import (
+    NODATA_IDX,
+    NODATA_RGBA,
+    TRANSPARENT_IDX,
+    TRANSPARENT_RGBA,
+    get_color_palette,
 )
+from harmony_browse_image_generator.exceptions import HyBIGError
 from harmony_browse_image_generator.sizes import (
+    GridParams,
     create_tiled_output_parameters,
     get_target_grid_parameters,
 )
@@ -49,23 +56,27 @@ def create_browse_imagery(
                                        driver=output_driver)
 
     try:
-        with rasterio.open(input_file_path, mode='r') as in_dataset:
+        with open_rasterio(
+            input_file_path, mode='r', mask_and_scale=True
+        ) as rio_in_array:
+            in_dataset = rio_in_array.rio._manager.acquire()
             validate_file_type(in_dataset)
 
-            if in_dataset.count == 1:
+            if rio_in_array.rio.count == 1:
                 color_palette = get_color_palette(
                     in_dataset, source, item_color_palette
                 )
-                raster = convert_singleband_to_raster(in_dataset, color_palette)
-            elif in_dataset.count in (3, 4):
-                raster = convert_mulitband_to_raster(in_dataset)
+                raster = convert_singleband_to_raster(rio_in_array, color_palette)
+            elif rio_in_array.rio.count in (3, 4):
+                raster = convert_mulitband_to_raster(rio_in_array)
             else:
                 raise HyBIGError(
-                    f'incorrect number of bands for image: {in_dataset.count}')
+                    f'incorrect number of bands for image: {rio_in_array.rio.count}'
+                )
 
             raster, color_map = prepare_raster_for_writing(raster, output_driver)
 
-            grid_parameters = get_target_grid_parameters(message, in_dataset)
+            grid_parameters = get_target_grid_parameters(message, rio_in_array)
             grid_parameter_list, tile_locators = create_tiled_output_parameters(
                 grid_parameters
             )
@@ -79,7 +90,7 @@ def create_browse_imagery(
                 logger.info(f'out image file: {tiled_out_image_file}: {tile_location}')
 
                 write_georaster_as_browse(
-                    in_dataset,
+                    rio_in_array,
                     raster,
                     color_map,
                     grid_parameters,
@@ -98,7 +109,7 @@ def create_browse_imagery(
     return processed_files
 
 
-def convert_mulitband_to_raster(dataset: DatasetReader) -> ndarray:
+def convert_mulitband_to_raster(data_array: DataArray) -> ndarray:
     """Convert multiband to a raster image.
 
     Reads the three/four bands from the file, then normalizes them to the range
@@ -106,26 +117,26 @@ def convert_mulitband_to_raster(dataset: DatasetReader) -> ndarray:
     just ensures that the output is 8bit.
 
     """
-    bands = dataset.read()
+    bands = data_array.to_numpy()
     norm = Normalize()
 
-    if dataset.count == 3:
+    if data_array.rio.count == 3:
         norm.autoscale(bands)
         raster = around(norm(bands) * 255.0).astype('uint8')
 
-    elif dataset.count == 4:
+    elif data_array.rio.count == 4:
         norm.autoscale(bands[0:3, :, :])
         partial_raster = around(norm(bands[0:3, :, :]) * 255.0).astype('uint8')
         raster = concatenate([partial_raster.data, bands[3:4, :, :]])
 
     else:
-        raise HyBIGError(f'Cannot create image from {dataset.count} band image')
+        raise HyBIGError(f'Cannot create image from {data_array.rio.count} band image')
 
     return raster
 
 
 def convert_singleband_to_raster(
-    dataset: DatasetReader,
+    data_array: DataArray,
     color_palette: ColorPalette | None = None,
 ) -> ndarray:
     """Convert input dataset to a 4 band raster image.
@@ -133,13 +144,13 @@ def convert_singleband_to_raster(
     Use a palette if provided otherwise return a greyscale image.
     """
     if color_palette is None:
-        return convert_gray_1band_to_raster(dataset)
-    return convert_paletted_1band_to_raster(dataset, color_palette)
+        return convert_gray_1band_to_raster(data_array)
+    return convert_paletted_1band_to_raster(data_array, color_palette)
 
 
-def convert_gray_1band_to_raster(dataset):
+def convert_gray_1band_to_raster(data_array: DataArray) -> ndarray:
     """Convert a 1-band raster without a color association."""
-    band = dataset.read(1)
+    band = data_array[0, :, :]
     cmap = matplotlib.colormaps['Greys_r']
     norm = Normalize()
     norm.autoscale(band)
@@ -152,10 +163,11 @@ def convert_gray_1band_to_raster(dataset):
     return reshape_as_raster(rgba_image[..., 0:3])
 
 
-def convert_paletted_1band_to_raster(dataset: DatasetReader,
-                                     palette: ColorPalette) -> ndarray:
+def convert_paletted_1band_to_raster(
+    data_array: DataArray, palette: ColorPalette
+) -> ndarray:
     """Convert a 1 band image with palette into a rgba raster image."""
-    band = dataset.read(1)
+    band = data_array[0, :, :]
     levels = list(palette.pal.keys())
     colors = [
         palette.color_to_color_entry(value, with_alpha=True)
@@ -164,12 +176,30 @@ def convert_paletted_1band_to_raster(dataset: DatasetReader,
     scaled_colors = [
         (r / 255.0, g / 255.0, b / 255.0, a / 255.0) for r, g, b, a in colors
     ]
-    cmap, norm = matplotlib.colors.from_levels_and_colors(levels,
-                                                          scaled_colors,
-                                                          extend='max')
+
+    cmap, norm = matplotlib.colors.from_levels_and_colors(
+        levels, scaled_colors, extend='max'
+    )
+
+    # handle no data values
+    if palette.ndv is not None:
+        nodata_colors = palette.color_to_color_entry(palette.ndv, with_alpha=True)
+        cmap.set_bad(
+            (
+                nodata_colors[0] / 255.0,
+                nodata_colors[1] / 255.0,
+                nodata_colors[2] / 255.0,
+                nodata_colors[3] / 255.0,
+            )
+        )
+
     scalar_map = matplotlib.cm.ScalarMappable(norm=norm, cmap=cmap)
-    scaled_raster = scalar_map.to_rgba(band, bytes=True)
-    return reshape_as_raster(scaled_raster)
+    rgba_image = np.zeros((*band.shape, 4), dtype='uint8')
+    for row_no in range(band.shape[0]):
+        rgba_image[row_no, :, :] = scalar_map.to_rgba(
+            np.ma.masked_invalid(band[row_no, :]), bytes=True
+        )
+    return reshape_as_raster(rgba_image)
 
 
 def image_driver(mime: str) -> str:
@@ -179,7 +209,9 @@ def image_driver(mime: str) -> str:
     return 'PNG'
 
 
-def prepare_raster_for_writing(raster: ndarray, driver: str) -> (ndarray, dict | None):
+def prepare_raster_for_writing(
+    raster: ndarray, driver: str
+) -> tuple[ndarray, dict | None]:
     """Remove alpha layer if writing a jpeg."""
     if driver == 'JPEG':
         if raster.shape[0] == 4:
@@ -189,7 +221,14 @@ def prepare_raster_for_writing(raster: ndarray, driver: str) -> (ndarray, dict |
     return palettize_raster(raster)
 
 
-def palettize_raster(raster: ndarray) -> (ndarray, dict):
+def remove_alpha(raster: ndarray) -> tuple[ndarray, ndarray | None]:
+    """Pull raster array off of input if it exists."""
+    if raster.shape[0] == 4:
+        return raster[0:3, :, :], raster[3, :, :]
+    return raster, None
+
+
+def palettize_raster(raster: ndarray) -> tuple[ndarray, dict]:
     """convert an RGB or RGBA image into a 1band image and palette.
 
     Converts a 3 or 4 band np raster into a PIL image.
@@ -200,23 +239,34 @@ def palettize_raster(raster: ndarray) -> (ndarray, dict):
     treated as either transparent or opaque and any transparent values are
     written to the final raster as 254 and add the mapped RGBA value to the
     color palette.
-
     """
-    # 0 to 253;
-    # reserve 254 for transparent images
-    # reserve 255 for off grid fill values
+    # reserves 254 for transparent images and 255 for off grid fill values
+    # 0 to 253
     max_colors = 254
+    rgb_raster, alpha = remove_alpha(raster)
 
-    # TODO [MHS, 11/01/2023] DAS-2020 raster, alpha = remove_alpha(raster)
-    multiband_image = Image.fromarray(reshape_as_image(raster))
+    multiband_image = Image.fromarray(reshape_as_image(rgb_raster))
     quantized_image = multiband_image.quantize(colors=max_colors)
 
     color_map = get_color_map_from_image(quantized_image)
-    # TODO [MHS, 11/01/2023] DAS-2020 if transparency replace values in
-    # quantized image and color_map
 
-    one_band_raster = np.expand_dims(np.array(quantized_image), 0)
+    quantized_array, color_map = add_alpha(alpha, np.array(quantized_image), color_map)
+
+    one_band_raster = np.expand_dims(quantized_array, 0)
     return one_band_raster, color_map
+
+
+def add_alpha(
+    alpha: ndarray | None, quantized_array: ndarray, color_map: dict
+) -> tuple[ndarray, dict]:
+    """If the input data had alpha values, manually set the quantized_image
+    index to the transparent index in those places."""
+    max_alpha = 255
+    if alpha is not None and np.any(alpha != max_alpha):
+        # Set any alpha to the transparent index value
+        quantized_array = np.where(alpha != max_alpha, TRANSPARENT_IDX, quantized_array)
+        color_map[TRANSPARENT_IDX] = TRANSPARENT_RGBA
+    return quantized_array, color_map
 
 
 def get_color_map_from_image(image: Image) -> dict:
@@ -246,7 +296,6 @@ def get_tiled_filename(input_file: Path, locator: dict | None = None) -> Path:
     Only update if there is a valid locator dict.
     """
     if locator is not None:
-        print(f'locator: {locator}')
         return input_file.with_suffix(
             f".r{int(locator['row']):02d}c{int(locator['col']):02d}{input_file.suffix}")
     return input_file
@@ -280,21 +329,23 @@ def validate_file_type(dsr: DatasetReader) -> None:
         raise HyBIGError(f'Input file type not supported: {dsr.driver}')
 
 
-def get_destination(grid_parameters: dict, n_bands: int) -> ndarray:
+def get_destination(grid_parameters: GridParams, n_bands: int) -> ndarray:
     """Initialize an array for writing an output raster."""
     return np.zeros(
         (n_bands, grid_parameters['height'], grid_parameters['width']),
         dtype='uint8')
 
 
-def write_georaster_as_browse(dataset: DatasetReader,
-                              raster: ndarray,
-                              color_map: ndarray | None,
-                              grid_parameters: dict,
-                              driver='PNG',
-                              out_file_name='outfile.png',
-                              out_world_name='outfile.pgw',
-                              logger=Logger) -> None:
+def write_georaster_as_browse(
+    data_array: DataArray,
+    raster: ndarray,
+    color_map: dict | None,
+    grid_parameters: GridParams,
+    driver='PNG',
+    out_file_name='outfile.png',
+    out_world_name='outfile.pgw',
+    logger=Logger,
+) -> None:
     """Write raster data to output file.
 
     Writes the raster to an output file using metadata from the original
@@ -303,9 +354,9 @@ def write_georaster_as_browse(dataset: DatasetReader,
 
     """
     n_bands = raster.shape[0]
-    dst_nodata = 255
+    dst_nodata = NODATA_IDX
     if color_map is not None:
-        color_map[dst_nodata] = (0, 0, 0, 0)
+        color_map[dst_nodata] = NODATA_RGBA
 
     creation_options = {
         **grid_parameters,
@@ -320,14 +371,16 @@ def write_georaster_as_browse(dataset: DatasetReader,
 
     with rasterio.open(out_file_name, 'w', **creation_options) as dst_raster:
         for dim in range(0, n_bands):
-            reproject(source=raster[dim, :, :],
-                      destination=dest_array[dim, :, :],
-                      src_transform=dataset.transform,
-                      src_crs=dataset.crs,
-                      dst_transform=grid_parameters['transform'],
-                      dst_crs=grid_parameters['crs'],
-                      dst_nodata=dst_nodata,
-                      resampling=Resampling.nearest)
+            reproject(
+                source=raster[dim, :, :],
+                destination=dest_array[dim, :, :],
+                src_transform=data_array.rio.transform(),
+                src_crs=data_array.rio.crs,
+                dst_transform=grid_parameters['transform'],
+                dst_crs=grid_parameters['crs'],
+                dst_nodata=dst_nodata,
+                resampling=Resampling.nearest,
+            )
 
         dst_raster.write(dest_array)
         if color_map is not None: